commit 20190917

+++ miniproject/src/imageboard/bean/ImageBoardDTO.java
@@ -7,7 +7,7 @@ public class ImageBoardDTO {
 	private int seq;
 	private String imageId;
 	private String imageName;
-	private double imagePrice;
+	private int imagePrice;
 	private int imageQty;
 	private String imageContent;
 	private String image1;
